% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preconditioner.R
\name{as_preconditioner_inverse}
\alias{as_preconditioner_inverse}
\title{Treat an operator P ~ A^-1 as a preconditioner}
\usage{
as_preconditioner_inverse(P, ...)
}
\arguments{
\item{P}{A `linop` approximating `A^-1`.}

\item{...}{Passed to [preconditioner()].}
}
\value{
A `preconditioner`.
}
\description{
Applied by computing `z = P r`.
}
