% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize.R
\name{as_sparse}
\alias{as_sparse}
\title{Materialise an operator as a sparse matrix}
\usage{
as_sparse(A, max_entries = 1e+07)
}
\arguments{
\item{A}{A `linop`.}

\item{max_entries}{Guard on `nrow * ncol`.}
}
\value{
A `Matrix` object.
}
\description{
Materialise an operator as a sparse matrix
}
