% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialize.R
\name{explain}
\alias{explain}
\title{Report the structure and cost of an expression}
\usage{
explain(A)
}
\arguments{
\item{A}{A `linop`.}
}
\value{
Invisibly, a data frame of nodes.
}
\description{
Report the structure and cost of an expression
}
