% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linop.R
\name{linop_eye}
\alias{linop_eye}
\title{Identity operator}
\usage{
linop_eye(n, dtype = "double")
}
\arguments{
\item{n}{Dimension.}

\item{dtype}{`"double"` or `"complex"`.}
}
\value{
A `linop`.
}
\description{
Identity operator
}
