% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linop.R
\name{linop_scaling}
\alias{linop_scaling}
\title{Diagonal operator}
\usage{
linop_scaling(d, properties = NULL)
}
\arguments{
\item{d}{The diagonal.}

\item{properties}{Declared capabilities, as in [linop()].}
}
\value{
A `linop`.
}
\description{
Diagonal operator
}
