% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{provenance_refine}
\alias{provenance_refine}
\title{Refine a discretisation}
\usage{
provenance_refine(p, n_new, ...)
}
\arguments{
\item{p}{A provenance envelope.}

\item{n_new}{The new discretisation size.}

\item{...}{Passed to methods.}
}
\value{
Provider-defined.
}
\description{
Refine a discretisation
}
