% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{strip_provenance}
\alias{strip_provenance}
\title{Remove the provenance envelope}
\usage{
strip_provenance(A)
}
\arguments{
\item{A}{A `linop`.}
}
\value{
A `linop`.
}
\description{
Remove the provenance envelope
}
