% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preconditioner.R
\name{as_preconditioner}
\alias{as_preconditioner}
\title{Treat an operator M ~ A as a preconditioner}
\usage{
as_preconditioner(M, solver, ...)
}
\arguments{
\item{M}{A `linop` approximating `A`.}

\item{solver}{`function(M, R)` performing the solve.}

\item{...}{Passed to [preconditioner()].}
}
\value{
A `preconditioner`.
}
\description{
Applied by solving `M z = r`.
}
