% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{provenance_summary}
\alias{provenance_summary}
\title{One-line description of a provenance envelope}
\usage{
provenance_summary(p, ...)
}
\arguments{
\item{p}{A provenance envelope.}

\item{...}{Passed to methods.}
}
\value{
A string.
}
\description{
One-line description of a provenance envelope
}
